requirements.txt
numpy>=1.26
scipy>=1.11
shapely>=2.0
matplotlib>=3.8
pytest>=8.0
pillow>=10.0
